Introduce the preliminary stub for sending gRPC commands to Spark Connect servers
This contains the initial tooling to process the Spark Connect `.proto` files
for executing simple queries. The current state of this code is _very_ simple
and uses a simple command line loop for sending queries back and forth.

There is still much to do in order to return something resembling dataframes for
further query.

Sponsored-by: Databricks Inc