Create the SparkConnect interface to provide a clean client API
The raw Tonic APIs are not very user-friendly so the `SparkConnec5` interface
provides a straightforward way to interact with the Spark Connect server. The
only function of consequence is `sql` right now which takes a query string and
sends it off to the server.

There is still a lot to do in order to provide a suitable CLI and Rust-based
API.

Sponsored-by: Databricks Inc