Structure the tree to allow for wasm support and other binaries
This makes it a little easier to start splitting out wasm support along with
support for the CLI and other crates. The current state of the wasm support is
basic and does not work property at present. There will be further investigation
necessary into whether spark-connect can natively speak grpc-web or be provided
with a simple proxy.

Sponsored-by: Databricks Inc